+++ TokenBase.h
@@ -27,7 +27,7 @@ const unsigned int IDENTIFIER_TOKEN = 5;
 class TokenBase : public RefCounted
 {
 public:
-	TokenBase(unsigned int char_index_, unsigned int type_) : char_index(char_index_), type(type_) {}
+	TokenBase(size_t char_index_, unsigned int type_) : char_index(char_index_), type(type_) {}
 	virtual ~TokenBase() {}
 
 	inline unsigned int getType() const { return type; }
@@ -42,7 +42,7 @@ public:
 	inline const std::string& getStringLiteralValue() const;
 	inline const std::string& getCharLiteralValue() const;
 
-	unsigned int char_index;
+	size_t char_index;
 private:
 	unsigned int type;
 };
@@ -51,7 +51,7 @@ private:
 class IdentifierToken : public TokenBase
 {
 public:
-	IdentifierToken(const std::string& x, unsigned int char_index) : TokenBase(char_index, IDENTIFIER_TOKEN), val(x) {}
+	IdentifierToken(const std::string& x, size_t char_index) : TokenBase(char_index, IDENTIFIER_TOKEN), val(x) {}
 	
 	std::string val;
 };
@@ -60,7 +60,7 @@ public:
 class FloatLiteralToken : public TokenBase
 {
 public:
-	FloatLiteralToken(double x, char suffix_, unsigned int char_index) : TokenBase(char_index, FLOAT_LITERAL_TOKEN), val(x), suffix(suffix_) {}
+	FloatLiteralToken(double x, char suffix_, size_t char_index) : TokenBase(char_index, FLOAT_LITERAL_TOKEN), val(x), suffix(suffix_) {}
 	
 	char suffix; // 'f' or 'd' or 0 if not present.
 	double val;
@@ -70,7 +70,7 @@ public:
 class IntLiteralToken : public TokenBase
 {
 public:
-	IntLiteralToken(int64 x, int num_bits_, bool is_signed_, unsigned int char_index) : TokenBase(char_index, INT_LITERAL_TOKEN), val(x), num_bits(num_bits_), is_signed(is_signed_) {}
+	IntLiteralToken(int64 x, int num_bits_, bool is_signed_, size_t char_index) : TokenBase(char_index, INT_LITERAL_TOKEN), val(x), num_bits(num_bits_), is_signed(is_signed_) {}
 	
 	int64 val;
 	int num_bits;
@@ -81,7 +81,7 @@ public:
 class BoolLiteralToken : public TokenBase
 {
 public:
-	BoolLiteralToken(bool x, unsigned int char_index) : TokenBase(char_index, BOOL_LITERAL_TOKEN), val(x) {}
+	BoolLiteralToken(bool x, size_t char_index) : TokenBase(char_index, BOOL_LITERAL_TOKEN), val(x) {}
 	
 	bool val;
 };
@@ -90,7 +90,7 @@ public:
 class StringLiteralToken : public TokenBase
 {
 public:
-	StringLiteralToken(const std::string& x, unsigned int char_index) : TokenBase(char_index, STRING_LITERAL_TOKEN), val(x) {}
+	StringLiteralToken(const std::string& x, size_t char_index) : TokenBase(char_index, STRING_LITERAL_TOKEN), val(x) {}
 	
 	std::string val;
 };
@@ -99,7 +99,7 @@ public:
 class CharLiteralToken : public TokenBase
 {
 public:
-	CharLiteralToken(const std::string& x, unsigned int char_index) : TokenBase(char_index, CHAR_LITERAL_TOKEN), val(x) {}
+	CharLiteralToken(const std::string& x, size_t char_index) : TokenBase(char_index, CHAR_LITERAL_TOKEN), val(x) {}
 	
 	std::string val;
 };
@@ -157,7 +157,7 @@ const unsigned int COMMA_TOKEN = 10;
 class COMMA_Token : public TokenBase
 {
 public:
-	COMMA_Token(unsigned int char_index) : TokenBase(char_index, COMMA_TOKEN) {}
+	COMMA_Token(size_t char_index) : TokenBase(char_index, COMMA_TOKEN) {}
 };
 
 
@@ -165,7 +165,7 @@ const unsigned int OPEN_PARENTHESIS_TOKEN = 11;
 class OPEN_PARENTHESIS_Token : public TokenBase
 {
 public:
-	OPEN_PARENTHESIS_Token(unsigned int char_index) : TokenBase(char_index, OPEN_PARENTHESIS_TOKEN) {}
+	OPEN_PARENTHESIS_Token(size_t char_index) : TokenBase(char_index, OPEN_PARENTHESIS_TOKEN) {}
 };
 
 
@@ -173,7 +173,7 @@ const unsigned int CLOSE_PARENTHESIS_TOKEN = 12;
 class CLOSE_PARENTHESIS_Token : public TokenBase
 {
 public:
-	CLOSE_PARENTHESIS_Token(unsigned int char_index) : TokenBase(char_index, CLOSE_PARENTHESIS_TOKEN) {}
+	CLOSE_PARENTHESIS_Token(size_t char_index) : TokenBase(char_index, CLOSE_PARENTHESIS_TOKEN) {}
 };
 
 
@@ -181,7 +181,7 @@ const unsigned int OPEN_BRACE_TOKEN = 13;
 class OPEN_BRACE_Token : public TokenBase
 {
 public:
-	OPEN_BRACE_Token(unsigned int char_index) : TokenBase(char_index, OPEN_BRACE_TOKEN) {}
+	OPEN_BRACE_Token(size_t char_index) : TokenBase(char_index, OPEN_BRACE_TOKEN) {}
 };
 
 
@@ -189,7 +189,7 @@ const unsigned int CLOSE_BRACE_TOKEN = 14;
 class CLOSE_BRACE_Token : public TokenBase
 {
 public:
-	CLOSE_BRACE_Token(unsigned int char_index) : TokenBase(char_index, CLOSE_BRACE_TOKEN) {}
+	CLOSE_BRACE_Token(size_t char_index) : TokenBase(char_index, CLOSE_BRACE_TOKEN) {}
 };
 
 
@@ -197,7 +197,7 @@ const unsigned int OPEN_SQUARE_BRACKET_TOKEN = 15;
 class OPEN_SQUARE_BRACKET_Token : public TokenBase
 {
 public:
-	OPEN_SQUARE_BRACKET_Token(unsigned int char_index) : TokenBase(char_index, OPEN_SQUARE_BRACKET_TOKEN) {}
+	OPEN_SQUARE_BRACKET_Token(size_t char_index) : TokenBase(char_index, OPEN_SQUARE_BRACKET_TOKEN) {}
 };
 
 
@@ -205,7 +205,7 @@ const unsigned int CLOSE_SQUARE_BRACKET_TOKEN = 16;
 class CLOSE_SQUARE_BRACKET_Token : public TokenBase
 {
 public:
-	CLOSE_SQUARE_BRACKET_Token(unsigned int char_index) : TokenBase(char_index, CLOSE_SQUARE_BRACKET_TOKEN) {}
+	CLOSE_SQUARE_BRACKET_Token(size_t char_index) : TokenBase(char_index, CLOSE_SQUARE_BRACKET_TOKEN) {}
 	std::string suffix;
 };
 
@@ -214,7 +214,7 @@ const unsigned int COLON_TOKEN = 17;
 class COLON_Token : public TokenBase
 {
 public:
-	COLON_Token(unsigned int char_index) : TokenBase(char_index, COLON_TOKEN) {}
+	COLON_Token(size_t char_index) : TokenBase(char_index, COLON_TOKEN) {}
 };
 
 
@@ -222,7 +222,7 @@ const unsigned int RIGHT_ARROW_TOKEN = 18;
 class RIGHT_ARROW_Token : public TokenBase
 {
 public:
-	RIGHT_ARROW_Token(unsigned int char_index) : TokenBase(char_index, RIGHT_ARROW_TOKEN) {}
+	RIGHT_ARROW_Token(size_t char_index) : TokenBase(char_index, RIGHT_ARROW_TOKEN) {}
 };
 
 
@@ -230,7 +230,7 @@ const unsigned int EQUALS_TOKEN = 19;
 class EQUALS_Token : public TokenBase
 {
 public:
-	EQUALS_Token(unsigned int char_index) : TokenBase(char_index, EQUALS_TOKEN) {}
+	EQUALS_Token(size_t char_index) : TokenBase(char_index, EQUALS_TOKEN) {}
 };
 
 
@@ -238,7 +238,7 @@ const unsigned int PLUS_TOKEN = 20;
 class PLUS_Token : public TokenBase
 {
 public:
-	PLUS_Token(unsigned int char_index) : TokenBase(char_index, PLUS_TOKEN) {}
+	PLUS_Token(size_t char_index) : TokenBase(char_index, PLUS_TOKEN) {}
 };
 
 
@@ -246,7 +246,7 @@ const unsigned int MINUS_TOKEN = 21;
 class MINUS_Token : public TokenBase
 {
 public:
-	MINUS_Token(unsigned int char_index) : TokenBase(char_index, MINUS_TOKEN) {}
+	MINUS_Token(size_t char_index) : TokenBase(char_index, MINUS_TOKEN) {}
 };
 
 
@@ -254,7 +254,7 @@ const unsigned int FORWARDS_SLASH_TOKEN = 22;
 class FORWARDS_SLASH_Token : public TokenBase
 {
 public:
-	FORWARDS_SLASH_Token(unsigned int char_index) : TokenBase(char_index, FORWARDS_SLASH_TOKEN) {}
+	FORWARDS_SLASH_Token(size_t char_index) : TokenBase(char_index, FORWARDS_SLASH_TOKEN) {}
 };
 
 
@@ -262,7 +262,7 @@ const unsigned int BACK_SLASH_TOKEN = 23;
 class BACK_SLASH_Token : public TokenBase
 {
 public:
-	BACK_SLASH_Token(unsigned int char_index) : TokenBase(char_index, BACK_SLASH_TOKEN) {}
+	BACK_SLASH_Token(size_t char_index) : TokenBase(char_index, BACK_SLASH_TOKEN) {}
 };
 
 
@@ -270,7 +270,7 @@ const unsigned int ASTERISK_TOKEN = 24;
 class ASTERISK_Token : public TokenBase
 {
 public:
-	ASTERISK_Token(unsigned int char_index) : TokenBase(char_index, ASTERISK_TOKEN) {}
+	ASTERISK_Token(size_t char_index) : TokenBase(char_index, ASTERISK_TOKEN) {}
 };
 
 
@@ -278,7 +278,7 @@ const unsigned int LEFT_ANGLE_BRACKET_TOKEN = 25;
 class LEFT_ANGLE_BRACKET_Token : public TokenBase
 {
 public:
-	LEFT_ANGLE_BRACKET_Token(unsigned int char_index) : TokenBase(char_index, LEFT_ANGLE_BRACKET_TOKEN) {}
+	LEFT_ANGLE_BRACKET_Token(size_t char_index) : TokenBase(char_index, LEFT_ANGLE_BRACKET_TOKEN) {}
 };
 
 
@@ -286,7 +286,7 @@ const unsigned int RIGHT_ANGLE_BRACKET_TOKEN = 26;
 class RIGHT_ANGLE_BRACKET_Token : public TokenBase
 {
 public:
-	RIGHT_ANGLE_BRACKET_Token(unsigned int char_index) : TokenBase(char_index, RIGHT_ANGLE_BRACKET_TOKEN) {}
+	RIGHT_ANGLE_BRACKET_Token(size_t char_index) : TokenBase(char_index, RIGHT_ANGLE_BRACKET_TOKEN) {}
 };
 
 
@@ -294,7 +294,7 @@ const unsigned int LESS_EQUAL_TOKEN = 27;
 class LESS_EQUAL_Token : public TokenBase
 {
 public:
-	LESS_EQUAL_Token(unsigned int char_index) : TokenBase(char_index, LESS_EQUAL_TOKEN) {}
+	LESS_EQUAL_Token(size_t char_index) : TokenBase(char_index, LESS_EQUAL_TOKEN) {}
 };
 
 
@@ -302,7 +302,7 @@ const unsigned int GREATER_EQUAL_TOKEN = 28;
 class GREATER_EQUAL_Token : public TokenBase
 {
 public:
-	GREATER_EQUAL_Token(unsigned int char_index) : TokenBase(char_index, GREATER_EQUAL_TOKEN) {}
+	GREATER_EQUAL_Token(size_t char_index) : TokenBase(char_index, GREATER_EQUAL_TOKEN) {}
 };
 
 
@@ -310,7 +310,7 @@ const unsigned int DOUBLE_EQUALS_TOKEN = 29;
 class DOUBLE_EQUALS_Token : public TokenBase
 {
 public:
-	DOUBLE_EQUALS_Token(unsigned int char_index) : TokenBase(char_index, DOUBLE_EQUALS_TOKEN) {}
+	DOUBLE_EQUALS_Token(size_t char_index) : TokenBase(char_index, DOUBLE_EQUALS_TOKEN) {}
 };
 
 
@@ -318,7 +318,7 @@ const unsigned int NOT_EQUALS_TOKEN = 30;
 class NOT_EQUALS_Token : public TokenBase
 {
 public:
-	NOT_EQUALS_Token(unsigned int char_index) : TokenBase(char_index, NOT_EQUALS_TOKEN) {}
+	NOT_EQUALS_Token(size_t char_index) : TokenBase(char_index, NOT_EQUALS_TOKEN) {}
 };
 
 
@@ -326,7 +326,7 @@ const unsigned int AND_TOKEN = 31;
 class AND_Token : public TokenBase
 {
 public:
-	AND_Token(unsigned int char_index) : TokenBase(char_index, AND_TOKEN) {}
+	AND_Token(size_t char_index) : TokenBase(char_index, AND_TOKEN) {}
 };
 
 
@@ -334,7 +334,7 @@ const unsigned int OR_TOKEN = 32;
 class OR_Token : public TokenBase
 {
 public:
-	OR_Token(unsigned int char_index) : TokenBase(char_index, OR_TOKEN) {}
+	OR_Token(size_t char_index) : TokenBase(char_index, OR_TOKEN) {}
 };
 
 
@@ -342,7 +342,7 @@ const unsigned int EXCLAMATION_MARK_TOKEN = 33;
 class EXCLAMATION_MARK_Token : public TokenBase
 {
 public:
-	EXCLAMATION_MARK_Token(unsigned int char_index) : TokenBase(char_index, EXCLAMATION_MARK_TOKEN) {}
+	EXCLAMATION_MARK_Token(size_t char_index) : TokenBase(char_index, EXCLAMATION_MARK_TOKEN) {}
 };
 
 
@@ -350,7 +350,7 @@ const unsigned int DOT_TOKEN = 34;
 class DOT_Token : public TokenBase
 {
 public:
-	DOT_Token(unsigned int char_index) : TokenBase(char_index, DOT_TOKEN) {}
+	DOT_Token(size_t char_index) : TokenBase(char_index, DOT_TOKEN) {}
 };
 
 
@@ -358,7 +358,7 @@ const unsigned int QUESTION_MARK_TOKEN = 35;
 class QUESTION_MARK_Token : public TokenBase
 {
 public:
-	QUESTION_MARK_Token(unsigned int char_index) : TokenBase(char_index, QUESTION_MARK_TOKEN) {}
+	QUESTION_MARK_Token(size_t char_index) : TokenBase(char_index, QUESTION_MARK_TOKEN) {}
 };
 
 
@@ -366,7 +366,7 @@ const unsigned int BITWISE_AND_TOKEN = 36;
 class BITWISE_AND_Token : public TokenBase
 {
 public:
-	BITWISE_AND_Token(unsigned int char_index) : TokenBase(char_index, BITWISE_AND_TOKEN) {}
+	BITWISE_AND_Token(size_t char_index) : TokenBase(char_index, BITWISE_AND_TOKEN) {}
 };
 
 
@@ -374,7 +374,7 @@ const unsigned int BITWISE_OR_TOKEN = 37;
 class BITWISE_OR_Token : public TokenBase
 {
 public:
-	BITWISE_OR_Token(unsigned int char_index) : TokenBase(char_index, BITWISE_OR_TOKEN) {}
+	BITWISE_OR_Token(size_t char_index) : TokenBase(char_index, BITWISE_OR_TOKEN) {}
 };
 
 
@@ -382,7 +382,7 @@ const unsigned int BITWISE_XOR_TOKEN = 38;
 class BITWISE_XOR_Token : public TokenBase
 {
 public:
-	BITWISE_XOR_Token(unsigned int char_index) : TokenBase(char_index, BITWISE_XOR_TOKEN) {}
+	BITWISE_XOR_Token(size_t char_index) : TokenBase(char_index, BITWISE_XOR_TOKEN) {}
 };
 
 
@@ -390,7 +390,7 @@ const unsigned int LEFT_SHIFT_TOKEN = 39;
 class LEFT_SHIFT_Token : public TokenBase
 {
 public:
-	LEFT_SHIFT_Token(unsigned int char_index) : TokenBase(char_index, LEFT_SHIFT_TOKEN) {}
+	LEFT_SHIFT_Token(size_t char_index) : TokenBase(char_index, LEFT_SHIFT_TOKEN) {}
 };
 
 
@@ -398,12 +398,12 @@ const unsigned int RIGHT_SHIFT_TOKEN = 40;
 class RIGHT_SHIFT_Token : public TokenBase
 {
 public:
-	RIGHT_SHIFT_Token(unsigned int char_index) : TokenBase(char_index, RIGHT_SHIFT_TOKEN) {}
+	RIGHT_SHIFT_Token(size_t char_index) : TokenBase(char_index, RIGHT_SHIFT_TOKEN) {}
 };
 
 
 const std::string tokenName(unsigned int t);
-Reference<TokenBase> makeTokenObject(unsigned int token_type, unsigned int char_index);
+Reference<TokenBase> makeTokenObject(unsigned int token_type, size_t char_index);
 
 
 } // end namespace Winter
